edit data model after merge project structure branch
