+++ src/main/kotlin/data/remote/datasource/clothes/ClothesDataSource.kt
@@ -1,0 +1,4 @@
+package org.madrid.data.repositories.remote.datasource.clothes
+
+interface ClothesDataSource {
+}
