+++ src/main/kotlin/data/remote/datasource/weather/WeatherDataSource.kt
@@ -1,0 +1,5 @@
+package org.madrid.data.repositories.remote.datasource.weather
+
+
+interface WeatherDataSource {
+}
