+++ src/main/kotlin/data/remote/requestmodels/IpAddressRequestModel.kt
@@ -1,0 +1,5 @@
+package org.madrid.data.repositories.remote.requestmodels
+
+data class IpAddressRequestModel(
+    val ip: String,
+)
