+++ src/main/kotlin/data/remote/responsmodels/weatherModel/Daily.kt
