+++ src/main/kotlin/data/repositories/ClothesRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package org.madrid.data.repositories
+
+class ClothesRepositoryImpl {
+}
