+++ src/main/kotlin/data/repositories/LocationRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package org.madrid.data.repositories
+
+class LocationRepositoryImpl {
+}
