+++ src/main/kotlin/data/repositories/WeatherRepositoryImpl.kt
@@ -1,0 +1,4 @@
+package org.madrid.data.repositories
+
+class WeatherRepositoryImpl {
+}
