--- src/main/kotlin/data/repository/ClothesRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package org.madrid.data.repository
-
-class ClothesRepositoryImpl {
-}
