--- src/main/kotlin/data/repository/LocationRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package org.madrid.data.repository
-
-class LocationRepositoryImpl {
-}
