--- src/main/kotlin/data/repository/WeatherRepositoryImpl.kt
@@ -1,4 +1,0 @@
-package org.madrid.data.repository
-
-class WeatherRepositoryImpl {
-}
