--- src/main/kotlin/data/source/remote/dataSource/colthes/ClothesRemoteDataSource.kt
@@ -1,4 +1,0 @@
-package org.madrid.data.source.remote.dataSource.colthes
-
-class ClothesRemoteDataSource: ClothesDataSource {
-}
