--- src/main/kotlin/data/source/remote/dataSource/weather/WeatherDataSource.kt
@@ -1,5 +1,0 @@
-package org.madrid.data.source.remote.dataSource.weather
-
-
-interface WeatherDataSource {
-}
