--- src/main/kotlin/data/source/remote/dataSource/weather/WeatherRemoteDataSource.kt
@@ -1,4 +1,0 @@
-package org.madrid.data.source.remote.dataSource.weather
-
-class WeatherRemoteDataSource: WeatherDataSource {
-}
