--- src/main/kotlin/data/source/remote/requestModels/IpAddressRequestModel.kt
@@ -1,5 +1,0 @@
-package org.madrid.data.source.remote.requestModels
-
-data class IpAddressRequestModel(
-    val ip: String,
-)
